GitEndpointSelector, RegistryEndpointSelectorを共通コンポーネント化
